\name{stock04}
\alias{stock04}
\docType{data}
\title{log-returns of 9 stocks from 2004}
\description{Weekly log-returns of 9 stocks from 2004, analyzed in Yuan et al.
(2007)}
\usage{data(stock04)}
\format{
  The format is:
 num [1:52, 1:9] 0.002275 -0.003795 0.012845 0.017489 -0.000369 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:9] "Walmart" "Exxon" "GM" "Ford" ...
}
 
\source{
Yuan, M., Ekici, A., Lu, Z., and Monteiro, R. (2007). Dimension reduction and coefficient
estimation in multivariate linear regression. Journal of the Royal Statistical Society Series
B, 69(3):329--346.
}
\references{
Yuan, M., Ekici, A., Lu, Z., and Monteiro, R. (2007). Dimension reduction and coefficient
estimation in multivariate linear regression. Journal of the Royal Statistical Society Series
B, 69(3):329--346.
}

\keyword{datasets}
