\name{MRCE-package}
\alias{MRCE-package}
\alias{MRCE}
\docType{package}
\title{Multivariate regression with covariance estimation}
\description{
Computes the MRCE estimators (Rothman, Levina, and Zhu, 2010) 
and has the dataset \code{stock04} used in Rothman, Levina, and Zhu (2010),
originally analyzed in Yuan et al. (2007).  
}
\details{The primary function is \code{\link{mrce}}. The dataset  is  \code{\link{stock04}}.}
\author{Adam J. Rothman

Maintainer: Adam J. Rothman <arothman@umn.edu>
}
\references{
Rothman, A.J., Levina, E., and Zhu, J. (2010). Sparse multivariate regression with covariance estimation.
  Journal of Computational and Graphical Statistics 19:974--962.
  
Yuan, M., Ekici, A., Lu, Z., and Monteiro, R. (2007). Dimension reduction and coefficient
estimation in multivariate linear regression. Journal of the Royal Statistical Society Series
B 69(3):329--346.

Cho-Jui Hsieh, Matyas A. Sustik, Inderjit S. Dhillon, and Pradeep Ravikumar (2011).
Sparse inverse covariance matrix estimation using quadratic approximation. 
Advances in Neural Information Processing Systems 24, 2011, p. 2330--2338.


Jerome Friedman, Trevor Hastie, Robert Tibshirani (2010). 
Regularization Paths for Generalized Linear Models via Coordinate Descent. 
Journal of Statistical Software, 33(1), 1-22.
}

\keyword{package}
 